Generation - Base Program
Generation is the largest of the programs for mass computation of
nullifiable sets. It's just going to do the expansions from a specified
input record to a specified output record, both existing files.

I'm thinking that it wouldn't be that hard to generalize all this -
rather than hard-coding these for nullifiability, these could be used to
find sets that can('t) solve ANY target number, really. For Generation,
it's just a matter of not doing supersets. For the weeding program, the
exhaustive test could be generalized with just a few lines.

For a bit I was thinking maybe the weeding phase could be removed if
mutations could insert larger sets (solvable for whatever they replace),
but I don't think that's the case as you could still have to go through
values greater than M. I guess really you just have to increase M high
enough so that nothing's solvable to omit weeding, and that may be worth
it in some cases.

Here are the four programs 2.0 will have:
- Generation (expand from in to out)
- Weed (exhaustive test on in)
- Evaluate (view set representations from in)
- New (create blank record file)

And here are a few things I forgot to mention in the last TODO list, and
they concern Set Records:
- Resize a record, preserving the existing settings (could be useful for
  trimming record files down to one M-value each after running a larger
  case in the 64 G of memory I have)
- Enable User-Mode programs to add their own info header or something
  (maybe just don't use FSEEK absolute?)