Fixed Unused (Null) Record Imports
A few commits ago I added in some extra checks for null pointers in the
actual array of the record struct. But I inappropriately placed one on
the Import function, which meant that records couldn't be initially
allocated by it. Didn't realize. Now it's fixed. Also added another
couple of those.