Generation - Arguments and Record Importing
Handling command-line arguments as well as importing source and
destination records. I want to do error checking for the imports, but
it's a few cases for each, and I don't like the idea of copy-pasting it.
I'm thinking of having all the error stuff use a goto to the end of the
Main function, and implementing the result checking as either a separate
macro or function. So far this is program running as expected.