Generation - Open/Import Error Checking
Record files are opened and their contents imported through a function
now, which handles all the errors and prints all the user messages
needed. Now the Main function is cleaner, and I can focus on more
important things. Also, I added the threads argument. Still need to
figure out how to do the configuration for which expansion methods to
use.