Generation - Core Expansion Logic
Created the function that each thread will run. This function does both
expansion phases, as it's simpler, and will exit the whole process if
there's an error. The argument is used as a pointer to the thread
number, but maybe I'll change it to something simpler where it just
passes by value, essentially. Need to make the 'supplemental' functions,
as the current main program calls them, and also fill in the rest of the
Main function.