Set Record Marking Handles Unallocated Sets
Now the Mark function won't return the 'invalid input' error when it's
given a set that isn't allocated, and instead quietly just returns as if
the set's already marked. Now the error is always a reason to panic, and
you can safely assert that the return code isn't -2, no matter what's
allocated.