Source File for Shared Functions
I don't wanna copy-paste the same code for functions between all the
utils I'm making, so I'm putting them in a separate source file. It's to
be directly included in each program, not compiled as a separate object,
as it's not an interface, just some normal program functions. If I
wasn't making multiple programs, it'd be in the program source anyways.

Also, here are some other things I thought of that I should do for 2.0:
- Restructure codebase with single 'lib' and 'utils' directories
- Add License, Readme
- Clean up comments (there are too many in a lot of places)