Weed - Performing Testing, Finished
This now finishes off the Weed program. Same as before, I'm using VLAs
to store the threads and to provide them their numbers. Pretty much the
same. I do have to go back and handle errors with creating/joining
threads, as I noticed they didn't result in an exit, just an error
message.

That was fast, and also these commits are very early. Why? Because I'm
travelling today. Got nothing to do but focus on this. Unfortunately I
can't sign in to Slack to say "I'm on a plane" in #random.