Create - Whole Program
Wrote the Create program. It's got input validation to ensure the
M-Value Range is in the right order.

I think I have some testing and patching to do with these programs, but
they're pretty much done, for the most part. I'll now work on cleaning
stuff up, particularly the Set Record library and just comments in
general. Also, I'm also thinking of adding a sort-of wrapper script that
uses the utility programs to automatically do large generations. Not
sure how I want that to work though.