Command-Line Argument Parsing Function
I've written a general function for parsing command-line arguments for
any utility program. It's a variadic function, and it takes in an array
of symbols to determine each parameter type, as well as a count of
required parameters, the arguments themselves, and then continually
takes references to variables in which to store the processed arguments.

Right now, it doesn't handle options, but I'm thinking of implementing
that by having it accept a mutable string, and replace characters with
'\0' if they weren't given as an option. I can start using this function
in programs in the state it's in, and implement options later.