All Utils - Simplified Argument Parsing
Now all the utility programs are using this abbreviated argument
parsing. Very pleasing, and seems to be working fine by my tests.

I'm beginning to wonder if perhaps I should've made the common functions
into a separate object... I'm having to include some extra headers in
the programs that don't appear to be in use, and this variadic function
for just one case is kinda jank when you consider it as though it were
part of the program source. I dunno. I'll think about it.

I'm also gonna have to figure out something for more Record granularity,
else I won't be able to deal in length-8 sets. I could do the simple
solution of buffering records, having them load in a certain amount at
once, which would mean multiple passes of effectively the same duration.
Another option is to use shorter length Records, but have some number of
high values be coded in and remain constant. I could make a length-8
record which has the high value 147 baked in, and the M-value range be
124 to 128. It'd be effectively be length-7 124-128 but it'd function as
length-8 with that high value. Hmm, I like this idea a lot...