Command-Line Argument Option Handler
Fitted the argument parser to skip over a first argument being used for
passing in options, and created another function for capturing those
options for the program to use. It's another variadic function that
takes in booleans by reference, corresponding to characters in a string
of valid options. I don't like the idea of putting it all in a single
function, and this seems more flexible anyways.