Generation, Evaluate - Option Parser
Using the new option parser in these two utility programs. This means
the options are now at the front for Generation. Evaluate now has an
option for doing a short display, skipping the printing of set
representations. I might go into the other programs and put a call to
the parser there, just to invalidate any options passed.