Refactored Common to Interface Subsystem
Now instead of including a C source file into each program, it's to be
linked as a separate object. When I rearrange everything into a new
directory structure, this'll go in 'lib/'. Also went around the programs
and removed the now-unnecessary function declarations, as well as the
weird-looking includes. Also fixed a type mistake I'd not seen.