Generation - Option for Creating Blank Destination
Gave the program a command-line option to avoid importing the
destination record, and instead create a new record with the same
M-Value Range as the Source. I was thinking I could use it to pipe
different programs together, without having to create files, but the
library uses a seek, so that isn't yet possible. It's still a handy
feature. I think I'll replace the seek with repetitive read-ins.