Basic Script for Classic Program
Here I've made a script that does the same as the original main program,
finds all the innullifiable sets on a given search space from scratch by
repeated generations and weeding at the end, but this script uses the
new utilities. The filenames for the utils aren't standard (yet), so you
have to be sure and compile them each correctly. Also, since it uses the
utilities, it transitions between generations through file I/O, which is
a big slowdown on my HDD system. But ignoring that, it seems it performs
just about the same. Still though, you probably should look into that
and see if you can solve it with /dev/shm or memory filesystems.

Also, here are some things I thought of adding to the utils:
- Disable non-error messages to stderr by default, enable by -v
  'verbose' option
- Add the record information header message to other utils