Utils - Verbosity is Optional
For the utils Generation and Weed, which are actual processes, the
verbose messages are now off by default, and can be enabled with '-v'.
For the Evaluate and Create utils, I removed the unnecessary Import/
Export messages. Whenever it is that I implement progress checking, I'll
make an option for it as well.