Automatic Script - Memory Temp File
The Automatic Script (that's what I'm calling it now) uses a shared
memory file for its tempfile for storing the record between programs. It
uses /dev/shm, which works (on most systems), but it's maybe not the
best solution. But I don't care. It's dang faster than disk I/O on an
HDD. And memory usage doesn't really matter for this now at-a-glance
convenience script.

However, there seems to be a bit of a discrepancy between the runtimes
of this and the classic program for some of the 'benchmarking' cases I
like, and it's particularly pronounced on the N = 6, M <= 40 case. I've
looked into it a bit, and it only seems to happen in cases where there
are a lot of sets remaining that the weeding process has to deal with,
and further, this discrepancy only seems to exist in single-threaded
mode. So my hypothesis is that it's the difference in how multithreading
is handled: the classic program uses its main thread for multithreaded
operations, while the Weed program creates a new pthread for each, and
that slight extra bit of overhead might be responsible here. I'll maybe
test this, but I'm not gonna bother changing anything. Just interesting.