Exhaustive Test - Audit
Yeah, this code is pretty bad. Contrary to the comment I wrote, I
actually do have one idea for how to optimize it, and it's basically to
do a 'weak' pass that just skips a bunch of stuff that 'likely' won't go
anywhere, reducing the space enough that a full pass wouldn't be so
costly. For example, based on the way a generation is performed, there's
a good chance the very first computation involved in nullifying the set
will have to go above the max M-value in the source nullifiable sets
expanded from, since an unmarked set can't reduce to one of them.

In fact, in the context of making the complete database, where each
generation also gets a pass of the exhaustive test, marking every single
nullifiable set POSSIBLE, this optimization will actually catch
everything, no need for another pass. I tried this out by placing in a
manual guard, and it did in fact speed up considerably, and the output
appeared to match up perfectly. Actually a decent optimization that'll
save a lot of time in making the database.