Readme - Basic Overview
Giving a basic project overview and a condensed description of the
problem itself. I'll give a bit more elaboration on what makes it
interesting and then write about the different utilities. Maybe a little
bit about the inner workings, just like an explanation of Set Records,
but besides that not much is necessary.