Interface - Fixed Issue with Uninitialized Values
So when I was going through to ensure C17 compatibility, I was using the
C2x empty-initializer for this VLA, since I needed everything to be
zero-initialized. But of course C2x isn't out yet, so we can't have
that, so I whipped up this for-loop super quick to initialize the array,
thinking I couldn't screw it up. Well I did, and I didn't catch it in
static analysis as I'd done it BEFORE changing stuff for compatibility.
Anyways, now it's fixed. You better believe as soon as C2x comes out I'm
gonna use that empty-initializer, and remove all those '_Bool's and
annoying 'stdbool' includes.