Macro to Skip Validation
The Set Record, Superset, and Mutation libraries now make use of a
'NO_VALIDATE' macro, which when defined will skip any code for
validating input. This is useful for making a slight performance
increase when you know the program is correct and always gives valid
inputs, like in the case of my utilities. The performance difference is
quite minuscule, especially with compiler optimizations already enabled.