i HATE MY LIFEEE
Inserting this ONE LINE sped up the weeding process by at least 5-6x...
AND I COULD HAVE DONE THIS MONTHS AGO!!! This is such a meme, and I feel
so ridiculous for it.

Without this base case, the function spent a lot of time dealing with
size-2 sets and doing every possible operation, when in reality there's
only one thing to do with those: subtract, and see if it's zero. Yeah,
if you can figure out how to optimize stuff deep down there in small
sets, that'll speed things up massively, as in the end most of the
runtime of the function is spent with those smaller ones.

Yeah, it's annoying and all how I could've done this ages ago, But at
least now my code is super fast. The N = 6, M <= 40 case used to take
5.5s, but now it finishes in 1.5s. The weeding process itself sped up
from 4.7s to 0.77s. And for the N = 7, M <= 81 case, the user time went
from 11m23s to 1m57s (and these figures include a substantial amount of
constant disk I/O time)!

So now that my attention is on the Exhaustive Test program, what can I
do to improve it? I think I'll abstract away the recursive function and
make a separate user-level function for some extra checks just for the
initial set. That compile warning about 'max object size' is just
because the potential for a set to be length-0, so I'll handle that case
in there and ensure the recursive function only works with sets of
length 2 or larger. Perhaps I could also try to make an optimized case
for length-3 sets, since those'd be much simpler to work with, and like
I said, it's down with shorter sets that the function spends most of its
time. I could possibly even try to implement memoization at that level.
These are all things I can do to improve the existing code, but I still
think that in the long-term it'd be better to phase this out and go for
something more elegant, unless perhaps I can make this blazing fast.

I've been told to avoid recursion wherever possible, and I've definitely
come to understand why. Most of the optimizations I've made throughout
this project have been simply getting rid of recursion. However, it is
kinda necessary here. I guess a good rule of thumb for using recursion
would be to always make sure your base cases are solid and thorough. If
there's one you could add, even if your code will catch it later, just
do it anyways. This is gonna be saving me loads of runtime.