Exhaustive Test - Optimized Length-3
Now the recursive function bases out at length-3 sets, and a separate
function takes it from there. The function does a simple equality check,
and then tries out operations to see if it can make two values into the
remaining one. At first, I thought I had to do this in a for-loop to get
all the different possible arrangements, but as it turns out, I don't
have to, as anything can be rearranged by inverse operations to just the
original possibilities. How cool is that? I wonder if I could use that
principle to optimize the recursive logic. But anyways, this new
function is vastly simpler and will return a result much faster than
going back through the recursive function's logic.

Then I made another simplification: the user-mode function now also
ensures there are no zeros, so the other logic doesn't have to deal with
that. It'd be redundant to check if there are equal values every time
but then also check for zeros, as they can only come from equal values.

The N = 6, M <= 40 case now only takes 0.79s to run, and the weeding
phase of that has been reduced to 0.120s. As for the N = 7, M <= 81 case
(which, as I should have clarified earlier, was also just the weeding
phase), it went down to 25s. Also, it turns out disk I/O isn't included
in user time, meaning my measurements from before are also good. Holy
mackerel, this is so much faster than before!