Set Records - Basics for Progress Tracking
At the base level, progress tracking is implemented by providing the
query function with an object pointer and a number X, where every X sets
it'll update with the remaining sets to query. This is just in the
library's own function, and not at user-level.

The big idea I'm working with for progress tracking is that essentially
the Generation & Weed programs will implement a signal handler that
adds up the latest progress updates and outputs a summary (printing?
through a pipe? not sure). The Automatic Script can make use of another
program to repeatedly send the signal and draw a progress bar or
whatever, and for database generation, it can just be dealt with
manually.