Generation - Signal Blocking & Handling
I've figured out all my signal stuff. It starts out by blocking the
signal, so it won't come up for any of the threads I care about, and
then after the work threads start, it'll launch another thread that
unblocks it and just waits, so it'll be the only thread that catches the
signal. The handler right now just prints out a message, but what it'll
do is sum together all the progress data and output it to a named pipe
(given just as a command argument), and probably just in raw form since
handlers can't call library functions.