Set Records - Progress Counts Up
Switched the Record query progress reports to count total number of sets
starting from zero and going up. Mostly this is so that progress
variables can be initialized to zero and be accurate. Also implemented
some very crude progress reporting in the Generation program. It's not
safe, and I'll change it so that it uses 'write' instead of stdio, but
for now this proves that it works.