Generation - Variable Naming
Applied the same naming changes from the previous commit, and made the
vector of progress variables 'volatile'.

You know what would be nice? If the progress stuff could be consistent,
so I could have all the progress code abstracted away in the interface
subsystem or something. I'd have to make the Generation program do only
a single pass, not one for each expansion phase. I might change the Set
Record 'query' function into a 'scan' function, and while it'll still
output based on a bitmask, it'll output the entire bit-field stored by
the record to the function, so I can have a function that takes in the
bit-field and performs the mutation conditionally, essentially. I also
could just get rid of the size parameter for the output function, since
the output size is always known. Yeah, I could revamp the interface for
the Record Library a lot. I'll look into that as I'm also working on the
unified expansion program.