Interface - Error Checking/Handling Macro
Here's a macro that automatically does the error checking for system
faults and prints out an actually maybe helpful error message so you
don't have to worry about writing them. And here I've implemented it in
the Generation program as well to check that it works, which it does.

So, I don't think I'm going to implement a progress bar for the
automatic script yet. The progress tracking scheme isn't consistent,
what with the Generation program doing two separate queries and having
to output two different progresses. I'm also wanting to try and abstract
away as much of the progress checking code as possible. The solution I'm
looking at would be to change the Set Record 'query' function so that it
also outputs the bit-field with each set. This way, the Generation
program can perform only one query, having a function conditionally
perform the mutation based on the bit-field.

The next major change I'm gonna be working on is merging the expansion
programs, so I figure during that time I can also work on changing up
the Record library interface. There are many simplifications I could
make, I feel. I'll put a tag after I optimized the Exhaustive Test, but
not for progress tracking. That'll come with the library changes. Maybe
though I could just do the query stuff and get a stable version with
progress tracking first.