Utils - Simple Error Checking
Now using the macro for error checking in all the utilities. I hated
those 'perror' calls with custom error messages for each. Too many
unnecessary lines. It's understandable, sure, but it makes things like
that code-block where all the working threads are launched/joined harder
to parse, especially now that it has to launch a progress tracking
thread as well.

So I think yes, I'm going to deal with the Set Record Library interface
so I can make the progress tracking consistent and possibly abstract as
much of it away as I can. Then I'll get that progress bars squared away
in the automatic script, and I'll have a stable version with progress
tracking. After that I'll do the expansion program merge.