Set Records - Output Functions Pass Bit-Field
Now the output functions used for Query have a parameter to which the
actual bit-field recorded is passed. So sets will be output based on
whether certain bits are set a certain way, and for good measure the
actual bit-field will be output alongside it in case further processing
might be necessary, or the same function is being used.

So I'm going to try using this to simplify the Generation program down
to one Query, so the progress tracking is simplified and standardized
between both programs. I'm thinking (hoping) it won't negatively impact
performance. I don't really see how it could, and if anything I'm
thinking it'd improve due to the record only being scanned once. But
I'll just have to see.