Generation - Single Query, Single Progress Output
Now the Generation program uses that single query and outputs that
standard progress data. It seems like I was right about performance: a
small improvement.

So I'll see what I can do in the way of abstracting away the progress
stuff in the Interface subsystem (or something else), as well as
progress bars or something for the automatic script, and then after I'm
done with that I can go ahead and merge the expansion programs.