Interface - Progress Pusher, Filename Errors
Created a function for pushing a progress update as part of the signal
handler, so the procedure is standard between programs. I'll go into the
programs and replace their stuff with this, simplifying the code for
progress tracking for each.

Also, with the Generation program in particular, I figured it'd be
helpful to output also the name of the file that caused an error. This
is probably the only error the user really needs: everything else would
need a debug or something.