Interface - Macros Swallowing the Semicolon
The function-like macros I made for error checking can't really be used
as the statements in an if-else statement, as they expand directly to a
compound statement and the semicolon you'd typically place after would
make a null-statement and screw up the 'else' bit. This is the standard
idiom for 'swallowing the semicolon', which allows them to be used in
this fashion.