Automatic Script - Progress Tracking Display
Made it so that the Automatic Script requests and displays progress info
once a second. It took quite a bit of finagling and learning about the
shell to get all the signals and process nonsense to work out well.
Basically, I run the work program in the background, pass its PID to a
little loop for progress that's also backgrounded, wait for the work
program to end, then kill the loop.

One problem I had was that I was trying to get the PID of a process I'd
backgrounded which was the work program coupled with an exit on failure,
and that wouldn't work for signaling as the process was just another
shell instance, not the work program. Then also I had issues where I was
getting weird messages from bash talking about my signal, but the issue
there I think was that it was sending before my process had actually
properly started or something, as that got fixed by moving 'sleep' ahead
of 'kill'. Yeah, this was a lot of debugging. Why are shell scripts so
hard? I guess all the intricacies and builtins.

In the course of getting here, I had a difficult time doing the reading
binary numbers in from the named pipe, and I settled for a short Python
script, but I feel like I could maybe get 'od' to work and that might
give more flexibility, so I'll try that again.

I also have some more things to do to the script, like trapping SIGINT
so it can clean up all the files it makes and stuff (as well as doing
that when an error occurrs), and maybe renaming some variables and
getting it POSIX-compatible.