Expand - Validation, Supersets
Implemented a test for input set validation. Also went and basically
copied the existing Superset enumeration function, which works by just
iterating through the potential insertion values, advancing to the next
set element when it reaches that value.

I'm thinking maybe I could do something like that for inserting
equivalent pairs too, at least for the additives (and when I say that I
mean both addition and subtraction), since they are nice and ordered,
and it'd just be a case of having there be two running indices, if that
makes sense. For the multiplicatives I may just have it manually insert
stuff.