Expand - Basic Mutation, Patched Issues
Here's what I'm thinking for mutations in the new Expand program. The
logic for additive mutations will be just like the superset expansion,
just with two insertions to keep track of. I've implemented this for
addition, and with just a few modifications it's just as good for
subtraction as well. There's a more high-level mutation function that'll
use this as well as whatever solution I make for multiplication and
division, probably just raw insertion.

Apart from that, I fixed a few compile errors and inconsistencies.
Nothing really to test yet, but soon I'll work on a test program.