Expand - Arbitrary Pair Insertion Function
Created the function for inserting an arbitrary equivalent pair into a
set, completing the multiplicative mutation function. I was able to
simplify a lot from the version in the Mutate program.

Next time I'll figure out the mode stuff, so I can use all this in the
main user-mode 'expand' function, so I can get to testing all this. I'll
just define some bitmask constants in the header file.