Expand - Fixed Issues
So there was an issue with the additive mutation function where it would
just keep looping on the first mutation point, and that was because I
wasn't passing in the mutation point correctly, as on the secondary
helper functions the mutation point is an index on the expanded set.

There was also an issue with supersets and just completely wrong sets
coming from multiplicative mutation. I just replaced my arbitrary pair
insertion function with the one from the Mutate program, because it's
just better. It also detects double values automatically. In addition
I'm having it manually not do equivalent pairs for 1, as none work.

By my tests it seems like everything's working fine now, but the real
test will come from the Generation program.