Generation - Switch to New Expand Program
Very easy to swap out. Things are running fine, but they're not as much
of a speed-up as I'd hoped. For smaller cases it even takes a little bit
longer, but for some of the larger cases (I've only really tested N = 7,
M <= 81) it's a bit faster actually. Gosh, I really hope this program is
correct.

As for the little bit about the progress arrays, they're still used, I
was just misremembering a bit. It was just the type that was
unnecessary, and I removed that.