Expand - Min/Max Operands
Implemented parameters for controlling the min/max operands for
generating equivalent pairs. Basically for the higher-level function to
use to control the M-range of the output.

I'm realizing that I was wrong about only needing to do this with
subtraction, as I was only considering for the potential of the M-value
being below the range. But there will be cases too where the M-value is
above, so I need to account for that as well. In that case, it'll be
addition pairs, and only on that largest value. Next I'll implement the
separate logic for checking the M-value and controlling the additive
operations.