Expand - Supers M-range Logic
Implemented some logic to keep sets from superset expansion inside the
M-range.

So for additives, sum equivalent pairs would have to be done when the
set is above the M-range. But it'd only work if there's only one value
poking out, I guess. I don't really need to worry about only doing the
top value in my logic, since the function takes in a minimum major
addend, so I just need to worry about that possible second one.

One thing I forgot to mention last time is that I changed stuff so that
mutations still outputs stuff that'd otherwise be a superset. This needs
to be done in case supersets were disabled. Hadn't thought of that.