Expand - Rethinking M-range Logic
So I realized the 'only one poking out' thing is actually completely
general to mutations, and when generalized it's relevant to everything
with expansion. Since values can't be removed, when there's a value
above the M-range, supersets can't work, and when there's two, mutations
can't work. So now this logic of checking for values 'poking out' is
done in the top-level function, and I redid the supersets function so it
doesn't deal with that and makes more sense.

This makes the logic a whole lot easier for the rest. For additive
mutations, just need to check if the M-value's above or below, then only
do the appropriate mutations to rectify the set. Same with
multiplicatives. Also for both, this will ensure that the case of a
value 'poking out' is dealt with.