Expand - Additive Mutations M-range Logic
It wasn't quite as simple as I thought. It still has to consider how
there might be a second value that could work as the M-value, and it
also has to be careful to not erase the M-value without replacing it
with something else in-range. I'm still struggling to fully explain and
think through the logic, but I'll try my best here:

So, when the set is already in-range, we can substitute difference pairs
anywhere, as well as sum pairs, provided they don't squash the M-value
without a replacement. When the set is below-range, there's nothing sum
pairs can do, as they only can make smaller values, so we do difference
pairs that insert a new in-range M-value. When the set is above-range,
there's likewise nothing difference pairs can do, so we do sum pairs,
but only on that single largest value, and again provided that we're
careful to include an in-range M-value at the end.

We should revisit this and try and write better comments explaining the
logic.