Expand - Multiplicative Mutation Supersets
A few commits ago I corrected the program so that expanding with only
mutations would include sets which'd also be supersets, as previously
they'd been omitted as "supersets already covers them." I also forgot to
mention this change in the commit message, instead writing about it in
the following one. But I forgot another thing: I forgot to modify the
insertion function so this'd also work with multiplicative mutations. So
here I can finally say the program works correctly.