Expand - Additive Mutations by Pair Insertion
Redid the additive mutations so they don't use the special method that
'reduces' memory operations. This way, a lot of the logic is more
explicit and upfront, so it's easy to tell what's going on at each stage
and what'll be counted. Seems like again the performance change is
minimal.

My plan now is to merge the multiplicative pair insertion function with
this, so they can all share logic for M-range stuff. Again, making it
much clearer what's going on.