Expand - Unified Mutation Logic
Brought the multiplicative mutation logic into the same function. Now
everything is ready to handle an M-range rather than just a maximum
M-value. It's a bit slower still, and I may soon look back on this and
see if I can get the 'fewer memory operations' method to work the way I
want. But for the moment this is fine.

Next, I'll bring M-range to user-level and test it out to see just how
much faster it is for working on a granule. Then finally to close off
this stable version I'll look at the Exhaustive Test and see what I can
do to add options and things. Also gotta look through the commit logs
because I've made some notes there for things to do as well.