Expand - Removed Efficient Additive Mutations
They're just too complex to deal with. They do in fact save memory
operations and runtime on larger cases, but there's just too much stuff
to bother with in the way of using them. For now, I'll stick to the
consistent option of using the arbitrary pair insertion, which isn't
that much of a penalty to runtime really. I'll maybe come back to this
if I really want performance, but it'd have to be general to all
operations and able to be abstracted away as a function independent of
finding the equivalent pairs (i.e. like the pair insertion function
now).

Here's another feature to add in the next version: you might want to
stop a generation midway because the output has few enough sets that
it'd be easier to just brute-force them, so add the number of remaining
sets in the output to progress tracking, and create a SIGINT handler to
optionally save midway. Also thinking a bit about database generation:
it'd be better to start with supersets because they're less expensive
(they use the smart method), and also because then you have the
supersets that don't need to be mutated anymore, which is more
information.