Exhaustive Test - Operation Order Optimization
So here I basically just switched around the order in which different
operations are done on a pair to reduce the set. It seems to make things
faster, and when I think about it, I feel like most of the time
subtraction has a better chance of getting a set to where it's easily
nullified. This sort of change will make it easy too to implement the
only-additive/multiplicative configuration.

One thing I'm thinking about now is that maybe I shouldn't be iterating
over pairs and then operations, but maybe operations and then pairs,
since I'm so sure that subtraction is just this much more likely to go
places. Like, multiplication? Last resort. So restructuring it this way
might make it even faster.