Interface, Generation, Weed - Output Progress
Added some space for progress information about the output, and using
them in the Generation and Weed programs. Generation gives the number of
unmarked sets remaining in the output, while Weed gives a count of the
number of sets that passed the exhaustive test so far. Need to figure
out how (if) to work this into the Automatic Script.

Now I'm thinking about maybe having another signal to dump the current
state of the output, but without quitting right after. Or maybe this
could be the same signal as progress updates, and it'd just be enabled
by a command-line flag. This could be useful for just backing stuff up
in case there's a crash or something.