Generation - Options for Progress Updates
Counting the number of yet-unmarked sets in the output record for every
single progress update is kinda a lot when you're requesting them five
times a second. So I made it optional. Also, I reworked the thing with
exporting the working state of the output so that it's an option for
progress updates instead, and that a progress update can be invoked on
the interrupt signal.