Exhaustive Test - Minimum First Computation
Implemented the option for restricting the first computation when
nullifying a set. In the database generation, this'll greatly optimize
weeding times, and it works because we know the range of all nullifiable
sets expanded into a record, and thus what any remaining sets can't
reduce to.

Here's another optimization I've been thinking about, and I don't think
I've written it down here: basically, avoid doing 'symmetric'
operations. Like, if you have {A, B, C, D}, you could operate on A and B
first, then C and D, but you could also operate on C and D first, then A
and B, and there'd be no difference. You should try and find a way to be
smart and avoid or reduce this. I was thinking earlier about having like
an index parameter that restricted operations solely to the left of it,
but I haven't thought that logic through. An idea, though.

So, now I'm getting at the set record extensions. I've been trying to
avoid these, because I just haven't had a 'perfect' idea of how I want
it to work. So I'm just going to implement it in the most basic way I
can, and only implement the features I think I need for the database
generation: expansion/weeding status, and fixed upper values. I'll just
have a mostly plaintext header that's user-mode so the programs can
implement this themselves. The library will scan its header at some
offset ahead of this 'reserved space'.

I'm not going to make any fancy logic or anything for the fixed upper
values. I'll just have something manually filter out sets that don't fit
with the record's M-range. It's just simpler, and probably there isn't
too much performance to be gained anyways. I can implement it on the
user-level too.