Interface - Basic Record Extensions Header
Basic implementation of Set Record Extensions using a plaintext header
with numbers. Read/written on import/export, and the values are given by
a little struct. If no values need to be specified, null-pointer may be
passed instead. Right now it's just the Complete Source M-range thing.

Next, I'm going to actually have the programs make use of the M-range.
The Weed program will write '-1' to records it deals with, the
Generation program will update with the source M-range if it's weeded
(and otherwise keep the same), and then the Weed program can use the
exhaustive test optimization.