Interface - Skipping to Fixed Values
I think the whole thing about storing the 'history' of the record in one
number in a little header to be used by all the programs was just a bad
idea. It's not precise, it only works in certain use-cases, and I've
been pulling my hair out trying to get myself to implement it. In my
mind, this is a detail better left to the actual program doing the mass
computation. It can keep track of what it's done to each record.

Something to note in terms of logic: so a thorough expansion (one that
uses both methods and completes) of a weeded record onto a destination
is equivalent to running the exhaustive test, but limiting the M-range
of whatever the first computation yielded. Very similar to what I was
doing with the 'first computation' restriction... it's just that this I
think is more precise. I'll go back and change that 'optimization' in
the Exhaustive Test to control M-value in this way. Not really much of
an optimization anymore, kinda just a normal procedure, and so I'll
probably work that into user-level as well with the Weed program. We can
imagine that a thorough weeding will have these processes (expansion,
exhaustive test) cover the complete range of computable values, in a
sense.

So I'll skip over this 'contiguous complete source' nonsense, and just
leave that to the mass computation setup. I'm starting work right now on
the fixed values, which is definitely something that needs to be
implemented on the program level. It's going to use the user-mode
headers, and I'll need to change the interface import function to
allocate the record information structure itself, since the record might
not be the expected size.